feat: Replace EmptyContent by Welcome component
